requirements.txt
geopy
ipython
jupyter
matplotlib
numpy
pandas
pyephem
pysolar
pytz
sunpy
